v0.3.3: uint8 -> ContentType

+++ protocol/file.go
@@ -93,8 +93,8 @@ type ImageContent interface {
 type AudioContent interface {
 	FileContent
 
-	Duration() int
-	SetDuration(duration int)
+	Duration() float64
+	SetDuration(duration float64)
 }
 
 /**
